feat: update to use add claims and optional required

+++ code/res/item.py
@@ -1,4 +1,4 @@
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from flask_restful import Resource, reqparse
 
 from code.models.item import ItemModel
@@ -31,7 +31,12 @@ class Item(Resource):
             return {"message": "An error while inserting item"}, 500
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims["is_admin"]:
+            return {"message": "Admin role required"}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_on_db()
@@ -51,5 +56,14 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=False)
     def get(self):
-        return {"items": [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = {"items": [item.json() for item in ItemModel.find_all()]}
+        if user_id:
+            return items
+        else:
+            return {"items": [item["name"] for item in items],
+                    "message": "More data available if you log in."}, 200
+
+
